Moved log.ts
- Moved log.ts to 'utils' directory.
